@@ -1,10 +1,10 @@
+
 source "http://rubygems.org"
 source "http://amee:aeC5ahx4@gems.amee.com"
-# Add dependencies required to use your gem here.
-# Example:
 
-gem "rails", "2.3.11"
+gem 'rails', '3.0.10'
 gem "amee", "~> 4.0.0"
-gem "quantify", "1.2.2"
+gem "quantify", "3.0.0"
 gem "google_visualr", "2.0.6"
+gem "sqlite3"
 
