@@ -1,13 +1,6 @@
 // Place your application-specific JavaScript functions and classes here
 // This file is automatically included by javascript_include_tag :defaults
 
-$(document).ajaxSend(function(event, request, settings) {
-  if (typeof(AUTH_TOKEN) == "undefined") return;
-  // settings.data is a serialized string like "foo=bar&baz=boink" (or null)
-  settings.data = settings.data || "";
-  settings.data += (settings.data ? "&" : "") + "authenticity_token=" + encodeURIComponent(AUTH_TOKEN);
-});
-
 function renderLoadingGif(wrappedSet) {
   var template = $('.ajaxloader');
   wrappedSet.each(function(){
@@ -18,9 +11,18 @@ function renderLoadingGif(wrappedSet) {
 $(document).ready(function(){
   $('#contentarea').hide().fadeIn('slow');
 
-  $("input[type='submit']").mouseenter(function() {
+  $("input[type='submit']").live('mouseenter',function() {
     $(this).css('cursor','pointer');
-    }, function() {
+  }).live('mouseleave', function() {
     $(this).css('cursor','auto');
   });
+
+  $(".selected-header").addClass('header-highlight');
+
+  $(".header-link:not(.selected-header)").live('mouseenter',function() {
+    $(this).addClass('header-highlight');
+  }).live('mouseleave', function() {
+    $(this).removeClass('header-highlight');
+  });
+
 });
